Break circular imports by making the boolean engines dynamic
